React implemented and the rest

+++ backend/controllers/goalControllers.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 // @desc    Get goals
 // @route   GET /api/goals
 // @acess   Private
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     
     res.status(200).json(goals)
 })
@@ -22,7 +23,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -39,6 +41,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal Not Found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    // Check for User
+    if (!user) {
+        res.status(401)
+        throw new Error('User Not Found')
+    }
+   
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not Authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
 
     res.status(200).json({msg: `Update goal ${updateGoal}`})
@@ -50,10 +66,23 @@ const updateGoal = asyncHandler(async (req, res) => {
 
 const deleteGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id)
-
+    
     if (!goal) {
         res.status(400)
         throw new Error('Goal Not Found')
+    }
+
+     const user = await User.findById(req.user.id)
+
+    // Check for User
+    if (!user) {
+        res.status(401)
+        throw new Error('User Not Found')
+    }
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not Authorized')
     }
     await Goal.deleteOne(goal)
     res.status(200).json({ msg: `Delete goal ${req.params.id}` })
